# docs/swagger.yaml
components:
  schemas:
    api.ErrorResponse:
      properties:
        error:
          example: validation failed
          type: string
      type: object
...
    api.createRunRequest:
      properties:
        dataset_id:
          type: string
        dataset_inline:
          additionalProperties: {}
          type: object
        duration_s:
          type: integer
        experiment_id:
          type: string
        labels:
          additionalProperties:
            type: string
          type: object
        meta_trace_id:
          type: string
        persona:
          type: string
        run_id:
          type: string
        vus:
          type: integer
        workflow_label:
          type: string
      type: object
    api.createRunResponse:
      properties:
        k6_job_name:
          type: string
        meta_trace_id:
          type: string
        run_id:
          type: string
        started_at:
          type: string
        status:
          type: string
      type: object
...
    run.Run:
      properties:
        dataset_id:
          type: string
        ended_at:
          type: string
        experiment_id:
          type: string
        id:
          type: string
        k6_job_name:
          type: string
        labels:
          additionalProperties:
            type: string
          type: object
        meta_trace_id:
          type: string
        reason:
          type: string
        started_at:
          type: string
        status:
          type: string
        workflow_id:
          type: string
        workflow_label:
          type: string
        workflow_name:
          type: string
      type: object
    stats.RunStats:
      properties:
        duration:
          $ref: '#/components/schemas/time.Duration'
        finalized_at:
          type: string
        iterations:
          type: integer
        latency_p50:
          $ref: '#/components/schemas/time.Duration'
        latency_p95:
          $ref: '#/components/schemas/time.Duration'
        latency_p99:
          $ref: '#/components/schemas/time.Duration'
        requests_dropped:
          type: integer
        requests_ok:
          type: integer
        requests_sent:
          type: integer
        run_id:
          type: string
        status:
          type: string
        variant_distribution:
          additionalProperties:
            type: integer
          type: object
        workflow_id:
          type: string
      type: object
...
    time.Duration:
      enum:
      - -9223372036854775808
      - 9223372036854775807
      - 1
      - 1000
      - 1000000
      - 1000000000
      - 60000000000
      - 3600000000000
      type: integer
      x-enum-varnames:
      - minDuration
      - maxDuration
      - Nanosecond
      - Microsecond
      - Millisecond
      - Second
      - Minute
      - Hour
...
info:
  description: 'Zeus is the execution plane for faults-lab. Manteion issues

    commands to it. This document is the integration contract between

    the two. Run lifecycle state machine and cross-cutting concerns

    (dataset handshake, error envelope) are documented in

    docs/api-contract.md.'
  title: Zeus Load Generation API
  version: '1.0'
openapi: 3.1.0
paths:
...
  /runs:
    get:
      description: 'Returns an envelope { "runs": [...] }. Filters are AND-combined;
        "since" must be

        RFC 3339 (invalid timestamps are silently ignored).'
      parameters:
      - description: filter by run status
        in: query
        name: status
        schema:
          type: string
      - description: filter by experiment id
        in: query
        name: experiment_id
        schema:
          type: string
      - description: filter by workflow id
        in: query
        name: workflow_id
        schema:
          type: string
      - description: RFC 3339 lower bound on started_at
        in: query
        name: since
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: envelope with runs array
      summary: List runs (cross-workflow)
      tags:
      - runs
  /runs/{run_id}:
    delete:
      description: 'Transitions a non-terminal run to "stopped". Refuses with 409
        when the run is

        already in a terminal state (completed/stopped/failed/rejected). On success,

        Prometheus per-run label series are pruned and SSE subscribers are closed.'
      parameters:
      - description: Run ID
        in: path
        name: run_id
        required: true
        schema:
          type: string
      responses:
        '204':
          description: run stopped
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: run not found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: run is already in terminal state
      summary: Stop run
      tags:
      - runs
    get:
      parameters:
      - description: Run ID
        in: path
        name: run_id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/run.Run'
          description: OK
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: run not found
      summary: Get run
      tags:
      - runs
  /runs/{run_id}/events:
    get:
      description: 'Server-Sent Events stream of run lifecycle and per-step events.
        Event types

        emitted: step.ok, step.drop, iteration.done, phase.transition. Used by manteion

        for live dashboards; not the canonical aggregated stats source. The Last-Event-ID

        header is reserved for future resumable-stream support; the current handler
        does

        not parse it (subscriber starts at the live edge).'
      parameters:
      - description: Run ID
        in: path
        name: run_id
        required: true
        schema:
          type: string
      - description: Last event ID for resumable stream (reserved; not yet honored)
        in: header
        name: Last-Event-ID
        schema:
          type: string
      responses:
        '200':
          content:
            text/event-stream:
              schema:
                type: string
          description: SSE stream of run events
        '404':
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: run not found
        '500':
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: streaming not supported by the underlying response writer
      summary: Stream run events (SSE)
      tags:
      - runs
  /runs/{run_id}/stats:
    get:
      description: 'Returns the finalized RunStats snapshot for a run. While the run
        is still in

        flight (or finished without a snapshot persisted), responds 200 with a minimal

        envelope { "run_id": ..., "status": "no stats available yet" } instead of
        404.

        The handler is tagged "metrics" rather than "runs" because it serves aggregate

        numbers, not run-document state.'
      parameters:
      - description: Run ID
        in: path
        name: run_id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/stats.RunStats'
          description: stats snapshot (or pending placeholder envelope)
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: run not found
      summary: Get run stats
      tags:
      - metrics
...
  /workflows/{id}/runs:
    get:
      description: 'Returns an envelope { "runs": [...] } of all runs (including terminal)
        bound

        to the workflow.'
      parameters:
      - description: Workflow ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: envelope with runs array
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: workflow not found
      summary: List runs for a workflow
      tags:
      - runs
    post:
      description: "Starts a run for the given workflow. Either dataset_id or dataset_inline\
        \ (capped\nat 1 MiB) may be supplied \u2014 they are mutually exclusive. When\
        \ the workflow\ndeclares a DataSchema, the dataset is validated; failures\
        \ produce a terminal\n\"rejected\" run and a 422 response with {run_id, status,\
        \ reason}. Other run-level\nrejections (workflow not found, run_id conflict,\
        \ oversized inline) return their\nnatural codes; the run is not persisted\
        \ in those cases. The success body is a\n202-Accepted envelope of run_id,\
        \ status, meta_trace_id, k6_job_name, started_at."
      parameters:
      - description: Workflow ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.createRunRequest'
                description: run request
                summary: body
        description: run request
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.createRunResponse'
          description: Accepted
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: invalid JSON, mutually exclusive dataset fields, or pool shape
            error
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: workflow or referenced dataset not found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: run_id conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: dataset_inline exceeds 1 MiB
        '422':
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: dataset failed schema validation; run created in rejected state
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: inline dataset persistence failure
      summary: Start workflow run
      tags:
      - runs
...
servers:
- description: Local dev (HTTP)
  url: http://localhost:8080/api/v1
- description: Local dev (HTTPS)
  url: https://localhost:8080/api/v1